Generate skeleton hooks file
Added `--generate-hooks` and `--template` options to **abao**.
The former generates a hooksfile from the given RAML and outputs
it to stdout. The latter allows specifcation of a user-provided
Mustache template file to be used by the generator.
Added related tests. Reorganized cmd-line unit tests.

Loosely based on PR from @galk-in

Fixes: #77